# api/static/openapi.yaml
openapi: 3.0.3
info:
  title: Asiga DB Access API
  description: |-
    WIP
  version: 0.0.1
servers:
  - url: https://127.0.0.1:5000/
tags:
  - name: chats
    description: chat history retrieval
paths:
  /chats/{oid}:
    get:
      tags:
        - chats
      summary: Retrieve chat history with the same id
      description: Retrieve single chat history
      parameters:
        - name: oid
          in: path
          description: ID of chat history to return
          required: true
          schema:
            type: string
            format: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chats'          
        '400':
          description: Parameters missing
        '404':
          description: objectID not found
  /chats:
    get:
      tags:
        - chats
      summary: Retrieve ALL chat messages from the database
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chats'    
  /chats/daterange/:
    get:
      tags:
        - chats
      summary: Retrieve chat messages from the database from a particular range in time
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chats'    
      parameters:
        - in: query
          name: from_date
          schema:
            type: string
            format: date
          description: start date
        - in: query
          name: to_date
          schema:
            type: string
            format: date
          description: start date
  /chats/date/:
    get:
      tags:
        - chats
      summary: Retrieve chat messages on the specified date
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chats'    
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
  /chats/summary/{word}:
    get:
      tags:
        - chats
      summary: Retrieve all chat messsages containing a specific word
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chats'    
      parameters:
      - name: word
        in: path
        description: Search for word in all chat summaries
        required: true
        schema:
          type: string
          format: string
components:
  schemas:
    chats:
      type: object
      properties:
        id:
          type: string
        summary:
          type: string
        date:
          type: string
          format: date-time
        chat_history:
          type: array
          items:
            $ref: '#/components/schemas/chat'
    chat:
      type: object
      properties:
        q_time:
          type: string
          format: date-time
        question:
          type: string
        answer:
          type: string
        a_time:
          type: string
          format: date-time
        sources:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 3